Refactor package structure to move open_ai and server components to internal directory
