+++ internal/bootstrap/llm.go
@@ -2,7 +2,7 @@ package bootstrap
 
 import (
 	"log"
-	"neko-acm/external/open_ai"
+	"neko-acm/internal/infrastructure/open_ai"
 )
 
 // 初始化大模型服务
