+++ internal/infrastructure/open_ai/model.go
