+++ internal/infrastructure/open_ai/openai.go
