+++ internal/infrastructure/open_ai/request.go
