+++ internal/infrastructure/open_ai/service.go
