+++ internal/infrastructure/open_ai/test.go
