+++ internal/interfaces/http/handler/misc.go
