+++ internal/service/chat/chat.go
@@ -2,7 +2,7 @@ package chat
 
 import (
 	"log"
-	"neko-acm/external/open_ai"
+	"neko-acm/internal/infrastructure/open_ai"
 	"neko-acm/internal/model"
 	"neko-acm/prompt"
 	"strings"
